Fix Vector2 to Vector3 conversion with explicit float casting
- Add explicit (float) casts to Vector2.X and Vector2.Y components
- Fix firstWaypoint.X/Y, fromPos.X/Y, toPos.X/Y, and targetPosition.X/Y conversions
- SharpDX.Vector3 constructor requires explicit float values, not Vector2 components
- This should resolve all remaining Vector type conversion errors

+++ AqueductBridgePlugin.cs
@@ -629,7 +629,7 @@ namespace AqueductBridge
                         if (currentPath.Count > 0)
                         {
                             var firstWaypoint = currentPath[0];
-                            var firstScreenPos = camera.WorldToScreen(new SharpDX.Vector3(firstWaypoint.X, firstWaypoint.Y, 0f));
+                            var firstScreenPos = camera.WorldToScreen(new SharpDX.Vector3((float)firstWaypoint.X, (float)firstWaypoint.Y, 0f));
                             
                             if (IsValidScreenPosition(playerScreenPos) && IsValidScreenPosition(firstScreenPos))
                             {
@@ -648,8 +648,8 @@ namespace AqueductBridge
                             var fromPos = currentPath[i];
                             var toPos = currentPath[i + 1];
                             
-                            var fromScreenPos = camera.WorldToScreen(new SharpDX.Vector3(fromPos.X, fromPos.Y, 0f));
-                            var toScreenPos = camera.WorldToScreen(new SharpDX.Vector3(toPos.X, toPos.Y, 0f));
+                            var fromScreenPos = camera.WorldToScreen(new SharpDX.Vector3((float)fromPos.X, (float)fromPos.Y, 0f));
+                            var toScreenPos = camera.WorldToScreen(new SharpDX.Vector3((float)toPos.X, (float)toPos.Y, 0f));
                             
                             if (IsValidScreenPosition(fromScreenPos) && IsValidScreenPosition(toScreenPos))
                             {
@@ -667,7 +667,7 @@ namespace AqueductBridge
                     if (Settings.ShowTargetMarker.Value && targetPosition.HasValue)
                     {
                         var camera = GameController.IngameState.Camera;
-                        var targetScreenPos = camera.WorldToScreen(new SharpDX.Vector3(targetPosition.Value.X, targetPosition.Value.Y, 0f));
+                        var targetScreenPos = camera.WorldToScreen(new SharpDX.Vector3((float)targetPosition.Value.X, (float)targetPosition.Value.Y, 0f));
                         
                         if (IsValidScreenPosition(targetScreenPos))
                         {
